frameworks/opt/net/wifi: wifi reconnect workaround
The networks validation in AutoJoin class relies on tsf value, expected
to be referenced to the machine boot time.
Some kernel drivers, at least in older kernel versions, do not support
this functionality and the framework generate a tsf referred to AP boot
seen time.
This cause the tsf to appear too old, as the AP was down, leading the
AutoJoin class to skip the AP.