nxp-mwifiex.git: patch: Reduce log level in wlan_fill_link_statistic
In the wlan_fill_link_statistic function, checking the maximum number
of supported channels, if exceeded, results in this warning message
continuing to print:
'wlan_fill_link_statistic: radio_stat->num_channels=1'.
This issue appeared starting with SDIW612 firmware version 18.99.2.p19.6
and also persists in latest versions.
The WiFi/BT chipset functionality is not affected in this case,
while waiting for the firmware to be improved, move this message
from error log level to warning to reduce verbosity.

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>