imx8mp-var-dart: Reduce wifi log verbosity
When using the default drvdbg value in the mwifiex module configuration,
the wifi module when scanning will print these messages continuously;

[ 1508.952081][  T565] wlan: wlan0 START SCAN
[ 1514.175488][   T74] wlan: SCAN COMPLETED: scanned AP count=5

this happens because this bit below is set(drvdbg=0x207):
bit 0:  MMSG            PRINTM(MMSG,...)
To reduce verbosity, set drvdbg to 0x206. The wifi module configuration
file, since it is modified, is kept in the device/variscite directory.